+++ app/src/main/java/com/example/spholder/proxy/SpRetrofit.java
@@ -5,7 +5,7 @@ import android.content.SharedPreferences;
 import android.text.TextUtils;
 
 import com.forjrking.preferences.crypt.Crypt;
-import com.forjrking.preferences.provide.ProvideKt;
+import com.forjrking.preferences.provide.ProvideKtKt;
 import com.forjrking.preferences.serialize.Serializer;
 
 import java.lang.reflect.InvocationHandler;
@@ -49,7 +49,7 @@ public final class SpRetrofit {
         if (!TextUtils.isEmpty(cryptKey) && !config.isMMKV()) {
 //            mCrypt = new InternalAESCrypt(cryptKey);
         }
-        SharedPreferences preferences = ProvideKt.createSharedPreferences(context, configName, cryptKey, config.isMultiProcess(), config.isMMKV());
+        SharedPreferences preferences = ProvideKtKt.createSharedPreferences(context, configName, cryptKey, config.isMultiProcess(), config.isMMKV());
         // 创建动态代理
         return (T) Proxy.newProxyInstance(cls.getClassLoader(), new Class<?>[]{cls}, new sharePreferencesProxy(preferences, mCrypt));
     }