+++ doc/CHANGELOG.md
@@ -6,14 +6,13 @@
 ### Fix
 
 ### TODO
-- 重命名bind方法名
-- 重构Binder
 - 多进程不能正常获取最新Value
 - default 擦除数据安全性
 - 拆分序列化支持库 gson moshi
 - jetpack publish 重命名 -ktx
 - 多进程exported=false
-- 使用AndroidKeyStore来实现AES加密(M+), 加密接口自定义支持
+- 使用AndroidKeyStore来实现AES加密(M+), 加密接口自定义支持  可以借助AI chatgpt
+- preferences androidx 实现性能评估
 
 ## UnRelease 2.0.0
 ### Added
@@ -22,6 +21,7 @@
 - 重构整体结构, 更改缓存为AtomicCache
 ### Remove
 - 重构Binder合并为一个类
+- 重命名bind方法名
 ### Fix
 - data class 修改属性不能正常apply
 - 多进程不能正常获取最新Value